Optimized pager loop (4-5 x faster on average)
Before we looped through all the items, which works, but with a lot of
pages it got too slow. This was caused by the itemAt() call, which gets
called on every usage of current(). So now we just loop through the
pages only by number and call the itemAt() only when we need to.